refactor: PrescriptionCard에서 ~State 분리
